Added support for multiple ports
Generic implementation

+++ lib-l6470dmx/include/l6470params.h
@@ -39,7 +39,7 @@ public:
 	void Dump(void);
 
 private:
-	bool isMaskSet(uint16_t) const;
+	bool isMaskSet(uint32_t) const;
 
 public:
     static void staticCallbackFunction(void *p, const char *s);
