+++ lib-l6470dmx/include/modeparams.h
@@ -44,7 +44,7 @@ public:
 	bool HasSwitch(void) const;
 
 private:
-	bool isMaskSet(uint16_t) const;
+	bool isMaskSet(uint32_t) const;
 
 public:
     static void staticCallbackFunction(void *p, const char *s);
